Implementa cadastro de Fazenda

+++ 2-cadastro-fazenda/src/main/java/com/caionastu/desafiojava/fazenda/api/FazendaController.java
@@ -3,19 +3,20 @@ package com.caionastu.desafiojava.fazenda.api;
 import com.caionastu.desafiojava.fazenda.dominio.Fazenda;
 import com.caionastu.desafiojava.fazenda.dominio.FazendaRepository;
 import com.caionastu.desafiojava.fazenda.excecoes.FazendaNaoEncontradaPeloIdException;
+import com.caionastu.desafiojava.fazenda.excecoes.FazendaNaoEncontradaPeloNomeException;
 import com.totvs.tjf.api.context.stereotype.ApiGuideline;
 import com.totvs.tjf.api.context.v2.request.ApiFieldRequest;
 import com.totvs.tjf.api.context.v2.request.ApiPageRequest;
 import com.totvs.tjf.api.context.v2.request.ApiSortRequest;
 import com.totvs.tjf.api.context.v2.response.ApiCollectionResponse;
+import com.totvs.tjf.core.api.jpa.repository.ApiJpaCollectionResult;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import jdk.javadoc.doclet.Reporter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Collection;
+import javax.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -30,33 +31,57 @@ public class FazendaController {
     private final FazendaRepository repository;
 
     @GetMapping
-    public ResponseEntity<ApiCollectionResponse<FazendaDTO>> buscarTodos(ApiPageRequest pageRequest, ApiSortRequest sortRequest, ApiFieldRequest fieldRequest) {
-        Collection<FazendaDTO> fazendas = repository.findAllProjected(fieldRequest, pageRequest, sortRequest)
+    @ApiOperation(value = "Busca todas as fazendas.")
+    public ApiCollectionResponse<FazendaDTO> buscarTodos(ApiPageRequest pageRequest, ApiSortRequest sortRequest, ApiFieldRequest fieldRequest) {
+        ApiJpaCollectionResult<Fazenda> fazendas = repository.findAllProjected(fieldRequest, pageRequest, sortRequest);
+        List<FazendaDTO> fazendasDTO = fazendas
                 .getItems()
                 .stream()
                 .map(FazendaDTO::from)
                 .collect(Collectors.toList());
 
-        return ResponseEntity.ok(ApiCollectionResponse.from(fazendas));
+        return ApiCollectionResponse.of(fazendasDTO, fazendas.hasNext());
     }
 
     @GetMapping(path = "/{id}")
+    @ApiOperation(value = "Busca a fazenda pelo id.")
     public ResponseEntity<FazendaDTO> buscarPeloId(@PathVariable UUID id) {
-        Fazenda fazenda = repository.findById(id).orElseThrow( () -> new FazendaNaoEncontradaPeloIdException(id));
+        Fazenda fazenda = repository.findById(id).orElseThrow(() -> new FazendaNaoEncontradaPeloIdException(id));
+        return ResponseEntity.ok(FazendaDTO.from(fazenda));
+    }
+
+    @GetMapping(path = "nome/{nome}")
+    @ApiOperation(value = "Busca a fazenda pelo nome.")
+    public ResponseEntity<FazendaDTO> buscarPeloNome(@PathVariable String nome) {
+        Fazenda fazenda = repository.findByNome(nome).orElseThrow(() -> new FazendaNaoEncontradaPeloNomeException(nome));
         return ResponseEntity.ok(FazendaDTO.from(fazenda));
     }
 
     @PostMapping
-    @ApiOperation(value = "Cria uma Fazenda.")
-    public List<Fazenda> criar(@RequestBody FazendaRequest request) {
+    @ApiOperation(value = "Cria uma fazenda.")
+    public ResponseEntity<FazendaDTO> criar(@Valid @RequestBody NovaFazendaRequest request) {
+        Fazenda fazenda = repository.save(request.toModel());
+        return ResponseEntity.ok(FazendaDTO.from(fazenda));
+    }
 
-        // TODO: 25/11/2020 Arrumar 
-        System.out.println("Fazenda Request");
-        return repository.findAll();
+    @PostMapping(path = "/{id}")
+    @ApiOperation(value = "Altera nome da fazenda.")
+    public ResponseEntity<FazendaDTO> alterar(@Valid @RequestBody AlteraFazendaRequest request, @PathVariable UUID id) {
+        Fazenda fazenda = repository.findById(id)
+                .orElseThrow(() -> new FazendaNaoEncontradaPeloIdException(id));
+
+        fazenda.alteraNome(request.getNome());
+        repository.save(fazenda);
+
+        return ResponseEntity.ok(FazendaDTO.from(fazenda));
     }
 
     @DeleteMapping(path = "/{id}")
-    public ResponseEntity<Void> remover(@PathVariable UUID id){
+    public ResponseEntity<Void> remover(@PathVariable UUID id) {
+        if (!repository.existsById(id)) {
+            throw new FazendaNaoEncontradaPeloIdException(id);
+        }
+
         repository.deleteById(id);
         return ResponseEntity.ok().build();
     }
